US3928: Move RestApiTable attribute
