+++ Gateway/MinistryPlatform.Models/Attributes/RestApiTable.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace MinistryPlatform.Models.Attributes
+{
+    [AttributeUsage(AttributeTargets.Class)]
+    public class RestApiTable : System.Attribute
+    {
+        public string Name { get; set; }
+    }
+}
